+++ smoothchange/src/com/rest/smoothchange/communication/plan/controller/CommunicationPlanController.java
@@ -175,10 +175,16 @@ public class CommunicationPlanController {
 	@ApiOperation(value = "Delete Communication Plans by Id")
 	@SuppressWarnings("unchecked")
 	@RequestMapping(value = "/DeleteCommunicationPlansById", method = RequestMethod.DELETE)
-	public ResponseEntity deleteCommunicationPlansById(@RequestHeader("API-KEY") String apiKey,@RequestParam("id") String id) throws UnauthorizedException {
+	public ResponseEntity deleteCommunicationPlansById(@RequestHeader("API-KEY") String apiKey,@RequestParam("id") String id) throws UnauthorizedException, NoRecordsFoundException {
 		if(!apiKey.equals(MessageEnum.API_KEY))
 		{
 			throw new UnauthorizedException(MessageEnum.unathorized);
+		}
+		CommunicationPlanDto dto=communicationPlanService.getById(Long.parseLong(id));
+		if(dto==null)
+		{
+			throw new NoRecordsFoundException(MessageEnum.enumMessage.ID_NOT_VALID.getMessage());
+
 		}
 		communicationPlanService.deleteById(Long.parseLong(id));
 		ResponseBean responseBean = new ResponseBean();
